Initial php to create JSON data from scraping sites
A function that takes two parameters (a site and number of desired
posts). The site is a name not a URL. The number of posts is just that.
A third parameter is optional, JSON yes or no. (not sure if going to
use)
Possibly feed the function an array based on input from the user.